fixes some entities and DbContext

+++ WebMarket.EF/Entities/BasketEntity.cs
@@ -5,9 +5,9 @@ namespace WebMarket.EF.Entities
 {
     public class BasketEntity
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public DateTime CreatedAt { get; set; }
