+++ WebMarket.EF/Entities/ProductCharacteristicEntity.cs
@@ -2,9 +2,9 @@
 {
     public class ProductCharacteristicEntity
     {
-        public int CharacteristicId { get; set; }
+        public long CharacteristicId { get; set; }
 
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public string Description { get; set; }
