+++ WebMarket.EF/Entities/SubCategoryEntity.cs
@@ -2,9 +2,9 @@
 {
     public class SubCategoryEntity
     {
-        public int SubCategoryId { get; set; }
+        public long SubCategoryId { get; set; }
 
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
 
         public string Description { get; set; }
